requirements.txt
lxml
python-decouple
apache-airflow-providers-apache-spark
requests
